requirements.txt
requests
pyfiglet
gmpy2
inflect
datetime
tabulate
pycryptodome
cryptography
termcolor
pytest